Using Pytorch Lightning - MLFlowlogger instead of try_mlflow_log - mlflow utility
